Add memory allocation to sorter.
Tha values that are grabbed from the stdin shpuld now have memory allocation and can be accessed later. There was also the addition of a new strtok_single function which now returns null if there are two adjacent delimiters. This can be used to set the position of a missing entry to "<EMPTY>".