Complete a fully implemented mergsort with no seg faults
At this point the mergesort correctly returns and ordering array which is then used to properly read values. This allows for minimal loading on the stack and heap. As of this commit there is more that needs to be done on the output end which involves outputting to a csv formatted file.